move proxy config to `consts`
