+++ packages/app/src/contexts/Proxies/index.tsx
@@ -14,7 +14,7 @@ import { ProxiesQuery } from 'api/query/proxiesQuery'
 import { AccountProxies } from 'api/subscribe/accountProxies'
 import BigNumber from 'bignumber.js'
 import type { AnyApi } from 'common-types'
-import { isSupportedProxy } from 'config/proxies'
+import { isSupportedProxy } from 'consts/util'
 import { useActiveAccounts } from 'contexts/ActiveAccounts'
 import { useApi } from 'contexts/Api'
 import { useExternalAccounts } from 'contexts/Connect/ExternalAccounts'
