+++ packages/app/src/hooks/useProxySupported/index.tsx
@@ -1,10 +1,8 @@
 // Copyright 2025 @polkadot-cloud/polkadot-staking-dashboard authors & contributors
 // SPDX-License-Identifier: GPL-3.0-only
 
-import {
-  UnsupportedIfUniqueController,
-  isSupportedProxyCall,
-} from 'config/proxies'
+import { UnsupportedIfUniqueController } from 'consts/proxies'
+import { isSupportedProxyCall } from 'consts/util'
 import { useActiveAccounts } from 'contexts/ActiveAccounts'
 import { useBonded } from 'contexts/Bonded'
 import { useProxies } from 'contexts/Proxies'
