+++ packages/app/src/overlay/modals/Accounts/Delegates/index.tsx
@@ -1,7 +1,7 @@
 // Copyright 2025 @polkadot-cloud/polkadot-staking-dashboard authors & contributors
 // SPDX-License-Identifier: GPL-3.0-only
 
-import { isSupportedProxy } from 'config/proxies'
+import { isSupportedProxy } from 'consts/util'
 import { useImportedAccounts } from 'contexts/Connect/ImportedAccounts'
 import type { ProxyDelegateWithBalance } from 'contexts/Proxies/types'
 import { AccountButton } from '../Account'
